fix ns error

+++ e2e/playwright/portal-client/database/seedMongo.js
@@ -7,15 +7,13 @@ const DB = 'mongo-e2e'
 
 async function dropCollection(mongoClient, collectionName) {
   const collection = mongoClient.db(DB).collection(collectionName)
-  // Check for existing collection before trying to drop it
-  if (collection) {
-    await collection.drop().catch((e) => {
-      console.log('error when dropping', e)
-      if (e.code !== 26) {
-        throw e
-      }
-    })
-  }
+
+  await collection.drop().catch((e) => {
+    console.log('error when dropping', e)
+    if (e.code !== 26) {
+      throw e
+    }
+  })
 }
 
 async function seedCollection(mongoClient, collectionName, jsonData) {
@@ -23,7 +21,7 @@ async function seedCollection(mongoClient, collectionName, jsonData) {
   await collection.insertOne(jsonData)
 }
 
-module.exports.seedDB = async () => {
+module.exports.seedDB = async (reset = true) => {
   // Connection URL
   const uri = `mongodb://localhost:27017`
 
@@ -35,7 +33,11 @@ module.exports.seedDB = async () => {
   try {
     await client.connect()
 
-    await dropCollection(client, 'users')
+    if (reset) {
+      await dropCollection(client, 'users')
+      console.log(`${DB} database reset!`)
+    }
+
     await seedCollection(client, 'users', testUser1)
     await seedCollection(client, 'users', testUser2)
